Re-organized all code
Re-organized all code through new project names, package names, condensed tests
