+++ README.md
@@ -1,1 +1,2 @@
-# CPS420_Project2
+# CPS420 Project 2
+Tests and Code for CPS 420 Project 2
