+++ src/tests/Loan.java
@@ -1,4 +1,4 @@
-package test;
+package tests;
 
 import java.time.LocalDate;
