+++ src/tests/Main.java
@@ -1,4 +1,4 @@
-package test;
+package tests;
 
 public class Main {
     public static void main(String[] args) {
