+++ src/tests/Member.java
@@ -1,4 +1,4 @@
-package test;
+package tests;
 
 import java.util.ArrayList;
 import java.util.List;
